+++ src/KorpiEngine.Runtime/Core/SceneManagement/Scene.cs
@@ -3,7 +3,7 @@ using KorpiEngine.Core.API.Rendering.Materials;
 using KorpiEngine.Core.API.Rendering.Shaders;
 using KorpiEngine.Core.EntityModel;
 using KorpiEngine.Core.EntityModel.Components;
-using KorpiEngine.Core.EntityModel.Systems.World;
+using KorpiEngine.Core.EntityModel.Systems;
 using KorpiEngine.Core.Rendering;
 using KorpiEngine.Core.Rendering.Cameras;
 using Entity = KorpiEngine.Core.EntityModel.Entity;
